% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDepth.R
\name{componentKDepth}
\alias{componentKDepth}
\title{Component \code{K}-depths for multivariate data}
\usage{
componentKDepth(X, K = 1, type = c("simplified", "full"))
}
\arguments{
\item{X}{The bivariate (or multivariate) dataset represented
as a numerical matrix with \code{p} columns and \code{n} rows.
The rows of \code{X} are typically the residuals.}

\item{K}{Positive integer, parameter of the depth.}

\item{type}{Type of the depth to compute. Can take values \code{simplified}
or \code{full}. By default, \code{simplified} is computed.}
}
\value{
The depth value; a numerical value between 0 and 1.
}
\description{
C++ implementations of the algorithm for
the computation of (both simplified and full) component
\code{K}-depth for multi-dimensional data (residuals).
The complexity of the algorithms is O(n).
}
\examples{
n = 500
p = 3
X = matrix(rnorm(p*n),ncol=p)
K = 2

componentKDepth(X,K,"simplified")*(n-K)
# the same evaluated component-wise
for(j in 1:p) print(componentKDepth(X[,j],K,"simplified")*(n-K))

componentKDepth(X,K,"full")*choose(n,K+1)
# the same evaluated component-wise
for(j in 1:p) print(componentKDepth(X[,j],K,"full")*choose(n,K+1))
}
\seealso{
\link{KSign} for the function called in the case of a full
version of the depth.
}
