% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDepth.R
\name{componentLDepth}
\alias{componentLDepth}
\title{Component \code{L}-depths for multivariate data}
\usage{
componentLDepth(X, L = 1, type = c("simplified", "full"))
}
\arguments{
\item{X}{The bivariate (or multivariate) dataset represented
as a numerical matrix with \code{p} columns and \code{n} rows.
The rows of \code{X} are typically the residuals.}

\item{L}{Positive integer, parameter of the depth.}

\item{type}{Type of the depth to compute. Can take values \code{simplified}
or \code{full}. By default, \code{simplified} is computed.}
}
\value{
The depth value; a numerical value between 0 and 1.
}
\description{
C++ implementations of the algorithm for
the computation of (both simplified and full) component
\code{L}-depth for multi-dimensional data (residuals).
The complexity of the algorithms is O(n).
}
\examples{
n = 500
p = 3
X = matrix(rnorm(p*n),ncol=p)
L = 2

componentLDepth(X,L,"simplified")*(n-L)
# the same evaluated component-wise
for(j in 1:p) print(componentLDepth(X[,j],L,"simplified")*(n-L))

componentLDepth(X,L,"full")*choose(n,L+1)
# the same evaluated component-wise
for(j in 1:p) print(componentLDepth(X[,j],L,"full")*choose(n,L+1))
}
\seealso{
\link{KSign} for the function called in the case of a full
version of the depth.
}
