% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDepth.R
\name{KSimplex}
\alias{KSimplex}
\title{Simplified and full \code{K}-simplex depth for bivariate data}
\usage{
KSimplex(X, K = 2, type = c("simplified", "full"))
}
\arguments{
\item{X}{The bivariate dataset represented as a numerical matrix
with two columns and \code{n} rows. The rows of \code{X} are
typically the residuals.}

\item{K}{Positive integer 1 or 2, parameter of the depth. If
\code{K} is 1, only the (simplified or full) 1-simplex depth is computed.
If \code{K} is 2, both (simplified or full) 1-simplex depth and (simplified
or full) 2-simplex depth are computed.}

\item{type}{Type of the depth to compute. Can take values \code{simplified}
or \code{full}.}
}
\value{
A vector of length \code{2}, both elements are numeric values
between 0 and 1. First element is the (simplified or full)
\code{1}-simplex depth. The second element it the (simplified or full)
\code{2}-simplex depth. If \code{K=1}, the second element of the vector is
set to be zero.
}
\description{
Wrapper function for the fastest C++ implementations of the algorithms for
the computation of the simplified \code{K}-simplex depths, and the
full \code{2}-simplex depth for two-dimensional data (residuals).
}
\examples{
n = 100
X = matrix(rnorm(2*n),ncol=2)
KSimplex(X, type="simpl")
KSimplex(X, type="full")
}
\seealso{
\link{simplKSimplex} and \link{fullKSimplex} for an implementation
of the simplified and full versions of the 1-simplex and 2-simplex depth,
respectively.
}
