% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDepth.R
\name{LSimplex_MC}
\alias{LSimplex_MC}
\title{A Monte Carlo approximation of the null distribution of
\code{L}-simplex depths for bivariate data}
\usage{
LSimplex_MC(n, B = 1e+05, full = TRUE, naive = NULL)
}
\arguments{
\item{n}{Sample size, positive integer.}

\item{B}{Number of independent Monte Carlo replicates to consider.
By default is set to 100000.}

\item{full}{Logical variable that says whether also the
full \code{L}-simplex depths should be considered. The procedure is
much faster without the full \code{L}-simplex depths.}
}
\value{
A large matrix of dimensions \code{B}-times-\code{j}, where
\code{j} is 2 if only the simplified \code{L}-simplex depths are considered
(\code{full=FALSE}), 4 if also the full \code{L}-simplex depths are
considered (\code{full=TRUE}): first column for the simplified
\code{1}-simplex depth,
second column for the simplified \code{2}-simplex depth, third column to the
full \code{1}-simplex depth (that is, the usual 2-dimensional simplicial
depth of the origin), and fourth column for the full \code{2}-simplex depth.
}
\description{
C++ implementation of a Monte Carlo approximation of the distribution
of the simplified \code{1}-simplex depth, simplified \code{2}-simplex depth,
and the full \code{1}-simplex depth and full \code{2}-simplex depth
for two-dimensional data (residuals).
}
\examples{
n = 100
B = 50
(LMC = LSimplex_MC(n,B))

# standardizing the columns to (approximately) pivotal quantities

mns = c(1/4, 1/12, 1/4, 1/12)
sds = c(1/4*sqrt(11/3)/sqrt(n-2),
        1/12*sqrt(169/10)/sqrt(n-3),
        1/n,
        1/n)

(LMC_stand = t((t(LMC)-mns)/sds)) # standardized kMC
}
\seealso{
\link{LSimplex}, \link{simplLSimplex}, and \link{fullLSimplex} for
the implementations of the functions used to generate the Monte Carlo
approximation.
}
