\name{SignDepth-package}
\alias{SignDepth-package}
\alias{SignDepth}
\docType{package}
\title{
  Sign depths for one-dimensional and bivariate data.
}
\description{
  Fast C++ and R implementation of the \code{K}-sign depth for univariate data, simplified 1-simplex depth and simplified 2-simplex depth for bivariate data, and the full 2-simplex depth for bivariate data.
}
\author{
Maintainer: Stanislav Nagy <nagy@karlin.mff.cuni.cz>
}
\references{
  Christine H. Muller, Stanislav Nagy, and Samuel Trippler (2025). Bivariate and multivariate sign depth and related distribution-free tests for model fit. \emph{Under review.}
}
\keyword{sign depth}
\keyword{\code{L}-simplex depth}
\keyword{\code{K}-sign depth}
